Update ssbapp.js
Added BFO, this BFO does not prevent the author of the idea from reducing pshiks. Huge respect to Jaffmer for the great work done and science for me in learning programming. I am a newbie on github, maybe I forgot something, pressed the wrong buttons. To use the latest versions, download the firmware ATS100_espruino.bin, the full version is 4 MB flash, it will work, that's for sure, this file contains the latest versions.